backend/requirements.txt
uvicorn>=0.32.0
websockets==14.2
fastapi>=0.115.2
requests>=2.32.3
python-multipart>=0.0.6
pinecone-client>=5.0.1
pinecone-plugin-assistant==1.1.0
pinecone-plugin-inference==1.0.3
pydantic>=2.5.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
aiohttp

annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.8
h11==0.14.0
idna==3.10
pinecone-client==5.0.1
pinecone-plugin-assistant==1.1.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
